+++ src/redux/WordsSelector.tsx
